Remove or address minor refactoring comments
